+++ src/lib/components/state/ZeroState.tsx
@@ -46,6 +46,11 @@ const ActionSection = ({
   );
 };
 
+/**
+ *
+ * @todo Will be refactored in the next PR
+ */
+
 export const ZeroState = ({
   list,
   isReadOnly,