+++ src/lib/pages/contract-list/components/ContractListTable.tsx
@@ -13,10 +13,12 @@ import {
   MenuList,
   MenuButton,
   MenuDivider,
+  chakra,
+  MenuItem,
 } from "@chakra-ui/react";
 import Link from "next/link";
 import {
-  MdMoreVert,
+  MdMoreHoriz,
   MdMode,
   MdOutlineBookmark,
   MdDelete,
@@ -25,7 +27,7 @@ import {
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import {
   AddToOtherList,
-  EditContract,
+  EditContractDetails,
   RemoveContract,
 } from "lib/components/modal/contract";
 import type { ContractInfo } from "lib/stores/contract";
@@ -34,11 +36,13 @@ import type { Option } from "lib/types";
 import { ContractNameCell } from "./table/ContractNameCell";
 import { TagsCell } from "./table/TagsCell";
 
-const iconProps = {
-  boxSize: "4",
-  display: "flex",
-  alignItems: "center",
-};
+const StyledIcon = chakra(Icon, {
+  baseStyle: {
+    boxSize: "4",
+    display: "flex",
+    alignItems: "center",
+  },
+});
 
 interface ContractListTableProps {
   contracts: ContractInfo[];
@@ -124,34 +128,29 @@ export const ContractListTable = ({
                       focusBorderColor="primary.main"
                       as={Button}
                     >
-                      <Icon
-                        as={MdMoreVert}
-                        style={iconProps}
+                      <StyledIcon
+                        as={MdMoreHoriz}
                         color="gray.600"
                         boxSize="6"
                       />
                     </MenuButton>
                     <MenuList>
-                      <EditContract
+                      <EditContractDetails
                         contractInfo={item}
-                        menuItemProps={{
-                          icon: (
-                            <Icon
-                              as={MdMode}
-                              style={iconProps}
-                              color="gray.600"
-                            />
-                          ),
-                          children: "Edit details",
-                        }}
+                        triggerElement={
+                          <MenuItem
+                            icon={<StyledIcon as={MdMode} color="gray.600" />}
+                          >
+                            Edit details
+                          </MenuItem>
+                        }
                       />
                       <AddToOtherList
                         contractInfo={item}
                         menuItemProps={{
                           icon: (
-                            <Icon
+                            <StyledIcon
                               as={MdOutlineBookmark}
-                              style={iconProps}
                               color="gray.600"
                             />
                           ),
@@ -163,14 +162,10 @@ export const ContractListTable = ({
                           <MenuDivider />
                           <RemoveContract
                             contractInfo={item}
-                            list={contractRemovalInfo}
+                            contractRemovalInfo={contractRemovalInfo}
                             menuItemProps={{
                               icon: (
-                                <Icon
-                                  as={MdDelete}
-                                  style={iconProps}
-                                  color="error.light"
-                                />
+                                <StyledIcon as={MdDelete} color="error.light" />
                               ),
                               children: "Remove from this list",
                             }}