+++ src/pages/contract-list/[slug].tsx
@@ -1,3 +1,3 @@
-import ContractsByList from "lib/pages/contracts/slug";
+import ContractsByList from "lib/pages/contract-list/slug";
 
 export default ContractsByList;